+++ spec/rspec/mocks/matchers/have_received_spec.rb
@@ -273,6 +273,47 @@ module RSpec
               expect(dbl).to have_received(:two).ordered
             }.to raise_error(/received :two out of order/m)
           end
+
+          context "when used with `with`" do
+            before do
+              dbl.one(1)
+              dbl.one(2)
+            end
+
+            it "passes when the order lines up" do
+              expect(dbl).to have_received(:one).with(1).ordered
+              expect(dbl).to have_received(:one).with(2).ordered
+            end
+
+            it "fails when the order is not matched" do
+              expect {
+                expect(dbl).to have_received(:one).with(2).ordered
+                expect(dbl).to have_received(:one).with(1).ordered
+              }.to fail_with(/received :one out of order/m)
+            end
+          end
+
+          context "when used on individually allowed messages" do
+            before do
+              allow(dbl).to receive(:foo)
+              allow(dbl).to receive(:bar)
+
+              dbl.foo
+              dbl.bar
+            end
+
+            it 'passes when the messages were received in order' do
+              expect(dbl).to have_received(:foo).ordered
+              expect(dbl).to have_received(:bar).ordered
+            end
+
+            it 'fails when the messages are received out of order' do
+              expect {
+                expect(dbl).to have_received(:bar).ordered
+                expect(dbl).to have_received(:foo).ordered
+              }.to raise_error(/received :foo out of order/m)
+            end
+          end
         end
       end